Use compatible dtype.

+++ pyramid_dit/pyramid_dit_for_video_gen_pipeline.py
@@ -248,7 +248,7 @@ class PyramidDiTForVideoGeneration:
         intermed_latents = []
 
         for i_s in range(len(stages)):
-            self.scheduler.set_timesteps(num_inference_steps[i_s], i_s, device=device)
+            self.scheduler.set_timesteps(num_inference_steps[i_s], i_s, device=device, dtype=dtype)
             timesteps = self.scheduler.timesteps
 
             if i_s > 0:
